# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "AdBoard API",
    "version": "v1"
  },
  "paths": {
...
    "/Auth/Login": {
      "post": {
        "tags": [
          "Authorization"
        ],
        "summary": "Аутентифицирует пользователя",
        "requestBody": {
          "description": "Модель для аутентификации",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginUserDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginUserDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/LoginUserDto"
              }
            }
          }
        },
        "responses": {
          "500": {
            "description": "Произошла внутрення ошибка",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "400": {
            "description": "Модель данных не валидна",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "200": {
            "description": "Запрос выполнен успешно"
          }
        }
      }
    },
    "/Auth/Register": {
      "post": {
        "tags": [
          "Authorization"
        ],
        "summary": "Создаёт пользователя",
        "requestBody": {
          "description": "Модель пользователя",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "500": {
            "description": "Произошла внутрення ошибка",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "400": {
            "description": "Модель данных не валидна",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "202": {
            "description": "Создание выполнено успешно"
          }
        }
      }
    },
    "/Categories": {
      "get": {
        "tags": [
          "Categories"
        ],
        "summary": "Вернуть все категории",
        "responses": {
          "500": {
            "description": "Произошла внутрення ошибка",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "200": {
            "description": "Запрос выполнен успешно",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ShortCategoryDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Categories"
        ],
        "summary": "Создаёт категорию по модели",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "required": [
                  "Name"
                ],
                "type": "object",
                "properties": {
                  "Name": {
                    "maxLength": 64,
                    "minLength": 3,
                    "type": "string",
                    "description": "Название"
                  },
                  "ParentId": {
                    "type": "string",
                    "description": "Id категории выше по иерархии (если есть)",
                    "format": "uuid"
                  }
                }
              },
              "encoding": {
                "Name": {
                  "style": "form"
                },
                "ParentId": {
                  "style": "form"
                }
              }
            }
          }
        },
        "responses": {
          "500": {
            "description": "Произошла внутрення ошибка",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "201": {
            "description": "Категория успешно создана",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShortCategoryDto"
                }
              }
            }
          },
          "400": {
            "description": "Модель данных запроса невалидна",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "Недостаточно прав для создания категории",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "422": {
            "description": "Произошёл конфликт бизнес логики",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          }
        }
      }
    },
    "/Categories/{id}": {
      "get": {
        "tags": [
          "Categories"
        ],
        "summary": "Получить категорию по id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Идентификатор категории",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Произошла внутрення ошибка",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "200": {
            "description": "Запрос выполнен успешно",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryDto"
                }
              }
            }
          },
          "404": {
            "description": "Категория с таким id не найдена",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Categories"
        ],
        "summary": "Обновляет категорию по модели",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Идентификатор обновляемой категории",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "required": [
                  "Name"
                ],
                "type": "object",
                "properties": {
                  "Name": {
                    "maxLength": 64,
                    "minLength": 3,
                    "type": "string",
                    "description": "Название"
                  },
                  "ParentId": {
                    "type": "string",
                    "description": "Id категории выше по иерархии (если есть)",
                    "format": "uuid"
                  }
                }
              },
              "encoding": {
                "Name": {
                  "style": "form"
                },
                "ParentId": {
                  "style": "form"
                }
              }
            }
          }
        },
        "responses": {
          "500": {
            "description": "Произошла внутрення ошибка",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "200": {
            "description": "Запрос успешно выполнен",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShortCategoryDto"
                }
              }
            }
          },
          "400": {
            "description": "Модель невалидна",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "Доступ запрещен",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "404": {
            "description": "Категория не найдена",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "422": {
            "description": "Конфликт бизнес логики",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Categories"
        ],
        "summary": "Удаляет категорию по ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Идентификатор категории",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Произошла внутрення ошибка",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "204": {
            "description": "Удаление произведено успешно"
          },
          "403": {
            "description": "Нет доступа",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "404": {
            "description": "Категория с таким ID не найдена",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          }
        }
      }
    },
    "/Categories/import": {
      "post": {
        "tags": [
          "Categories"
        ],
        "summary": "Импортирует категории из JSON списка.",
        "requestBody": {
          "description": "Коллекция категорий.",
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/ImportCategoryDto"
                }
              }
            },
            "text/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/ImportCategoryDto"
                }
              }
            },
            "application/*+json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/ImportCategoryDto"
                }
              }
            }
          }
        },
        "responses": {
          "500": {
            "description": "Произошла внутрення ошибка",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "200": {
            "description": "Категории успешно импортированы."
          },
          "422": {
            "description": "Произошёл конфликт бизнес логики.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CategoryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Id.",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "description": "Название.",
            "nullable": true
          },
          "parentId": {
            "type": "string",
            "description": "Id категории выше по иерархии.",
            "format": "uuid"
          },
          "childs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ShortCategoryDto"
            },
            "description": "Дочерние категории",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Данные о категории"
      },
...
      "CreateUserDto": {
        "required": [
          "login",
          "name",
          "password"
        ],
        "type": "object",
        "properties": {
          "name": {
            "maxLength": 64,
            "minLength": 4,
            "type": "string",
            "description": "Имя пользователя"
          },
          "login": {
            "maxLength": 64,
            "minLength": 4,
            "type": "string",
            "description": "Логин"
          },
          "password": {
            "maxLength": 50,
            "minLength": 8,
            "type": "string",
            "description": "Пароль"
          }
        },
        "additionalProperties": false,
        "description": "Данные для создания нового пользователя"
      },
      "ErrorDto": {
        "type": "object",
        "properties": {
          "errorCode": {
            "type": "string",
            "description": "Код ошибки.",
            "nullable": true
          },
          "userMessage": {
            "type": "string",
            "description": "Сообщение для пользователя.",
            "nullable": true
          },
          "internalErrors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ErrorDto"
            },
            "description": "Вложенные ошибки.",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Модель ошибки."
      },
      "ImportCategoryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "Идентификатор категории.",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "description": "Наименование категории.",
            "nullable": true
          },
          "parentId": {
            "type": "integer",
            "description": "Идентификатор родительской категории.",
            "format": "int32",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Модель категории для импорта из JSON."
      },
      "LoginUserDto": {
        "required": [
          "login",
          "password"
        ],
        "type": "object",
        "properties": {
          "login": {
            "maxLength": 64,
            "minLength": 4,
            "type": "string",
            "description": "Логин"
          },
          "password": {
            "maxLength": 50,
            "minLength": 8,
            "type": "string",
            "description": "Пароль"
          }
        },
        "additionalProperties": false,
        "description": "Информация о пользователе для аутентификации"
      },
...
      "ShortCategoryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Идентификатор",
            "format": "uuid"
          },
          "parentId": {
            "type": "string",
            "description": "Родительская категория",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "description": "Название",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Краткое описание категории"
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "apiKey",
        "description": "JWT Authorization header using the Bearer scheme.  \n                        Enter 'Bearer' [space] and then your token in the text input below.\n                        Example: 'Bearer secretKey'",
        "name": "Authorization",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}